Refactored into different files
