+++ src/model/Direction.ts
@@ -1,0 +1,14 @@
+/**
+ * @brief Represents a 3D direction. For all directions d, -d is the opposite
+ *        direction.
+ */
+enum Direction {
+  X_POSITIVE = 1,
+  X_NEGATIVE = -1,
+  Y_POSITIVE = 2,
+  Y_NEGATIVE = -2,
+  Z_POSITIVE = 3,
+  Z_NEGATIVE = -3,
+}
+
+export { Direction };
